+++ shell/BASH_SOURCE.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -x
+
+echo BASH_SOURCE=${BASH_SOURCE[@]}
+echo ROOT=$(dirname $(readlink -f ${BASH_SOURCE[0]}))